# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server
components:
  securitySchemes:
    securitySchemes:
      bearerAuth: # arbitrary name for the security scheme
    type: http
    scheme: bearer
    bearerFormat: JWT
...
paths:
...
  /updateSingleTask:
    post:
      tags:
        - tasks
      summary: Update single task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskTitle:
                  type: string
                taskDescription:
                  type: string
                assignedToID:
                  type: string
                labelColor:
                  type: string
                hoursExpected:
                  type: number
                hoursSpent:
                  type: number
      responses:
        '200':
          description: Task updated Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
        '500':
          description: something went wrong with updating task information
  /deleteSingleTask:
    delete:
      tags:
        - tasks
      summary: Delete single task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskID:
                  type: string
      responses:
        '200':
          description: Task deleted Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
        '500':
          description: something went wrong when deleting task
  /updateTaskPosition:
    post:
      tags:
        - tasks
      summary: Update task position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  _id:
                    type: string
                  position:
                    type: integer
      responses:
        '200':
          description: Task position successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
        '500':
          description: something went wrong when updating task position
  /updateTaskState:
    post:
      tags:
        - tasks
      summary: Update task state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskID:
                  type: string
                newStateID:
                  type: string
      responses:
        '200':
          description: Task state successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
        '500':
          description: something went wrong when updating task state
  /register:
    post:
      tags:
        - user
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                fName:
                  type: string
                lName:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Bad request
        '500':
          description: Internal Server Error
  /update-user:
    post:
      tags:
        - user
      summary: Update user information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fName:
                  type: string
                lName:
                  type: string
                color:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User updated successfully
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal Server Error
  /delete:
    delete:
      tags:
        - user
      summary: Delete user
      responses:
        '200':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /find-user:
    get:
      tags:
        - user
      summary: Find user by ID
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
                  email:
                    type: string
                  fName:
                    type: string
                  lName:
                    type: string
                  color:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal Server Error
...